+++ lib/scribe.ex
@@ -106,6 +106,30 @@ defmodule Scribe do
 
       iex> Scribe.print(data, colorize: false)
 
+  ### Text Alignment
+
+  Pass an `alignment` option of `:left`, `:center`, or `:right` for text alignment.
+  Defaults to `:left`.
+
+      iex> Scribe.print(data, alignment: :center)
+
+      +------------------------------+------------+--------+
+      |            :body             |  :current  |  :id   |
+      +------------------------------+------------+--------+
+      |   "A rather short string."   |    true    |  1234  |
+      |   "A rather short string."   |   false    |  2222  |
+      |   "A rather short string."   |    true    |  4444  |
+      +------------------------------+------------+--------+
+
+      iex> Scribe.print(data, alignment: :right)
+      +------------------------------+------------+--------+
+      |                        :body |   :current |    :id |
+      +------------------------------+------------+--------+
+      |     "A rather short string." |       true |   1234 |
+      |     "A rather short string." |      false |   2222 |
+      |     "A rather short string." |       true |   4444 |
+      +------------------------------+------------+--------+
+
   ### Styles
 
   Scribe supports five styling formats natively, with support for custom adapters.